+++ SOS/include/objects/player.h
@@ -11,9 +11,9 @@ class Player : public Entity {
 public:
     Player(Vec2 pos, SpriteData* spData, std::string objID);
     void setInput(PlayerInput* input) { this->input = input; }
-    void update(uint64_t deltaTime) override;
+    void update(float deltaTime) override;
     void accept(CollisionVisitor& visitor) override;
-    void handleInput(PlayerInput* input, uint64_t deltaTime);
+    void handleInput(PlayerInput* input, float deltaTime);
     void takeDamage(int amount);
     void collectItem();
     void applyPhysicsResponse(const Vec2& resolutionVector);
