+++ SOS/src/Object.cpp
@@ -14,7 +14,7 @@ Object::Object( Vec2 pos, ObjectType type, SpriteData* spData, std::string ID)
 }
 
 // Animation methods implementation
-void Object::updateAnimation(uint64_t deltaTime) {
+void Object::updateAnimation(float deltaTime) {
     animController.update(deltaTime);
 }
 
@@ -26,6 +26,17 @@ int Object::getCurrentSpriteIndex() const {
     return animController.getCurrentFrame();
 }
 
+/*
+* Adds an animation definition to the animation controller.
+* @param state The animation state to associate with this animation.
+* @param startFrame The starting frame index in the sprite sheet.
+* @param frameCount The number of frames in this animation.
+* @param framesPerRow The number of frames per row in the sprite sheet.
+* @param frameTime The time each frame is displayed in milliseconds.
+* @param loop Whether the animation should loop or not.
+* @note This function allows you to define multiple animations for different states of the object.
+*       For example, you can have different animations for walking, jumping, etc.
+*/
 void Object::addAnimation(AnimationState state, int startFrame, int frameCount, 
                          int framesPerRow, uint32_t frameTime, bool loop) {
     AnimationDef def(startFrame, frameCount, framesPerRow, frameTime, loop);
